[FP-73] Refactor: 피드백 적용
