+++ src/components/pages/myPage/Schedule/dateCell.tsx
@@ -1,4 +1,5 @@
 import Chip from '@/components/common/Chip';
+import { getCircleColor } from '@/utils/reservationCircle';
 import React from 'react';
 
 interface DateCellProps {
@@ -13,24 +14,6 @@ interface DateCellProps {
 }
 
 const DateCell: React.FC<DateCellProps> = ({ day, isCurrentMonth, reservation, onClick }) => {
-  const getCircleColor = () => {
-    if (reservation) {
-      if (reservation.completed > 0) {
-        return 'bg-gray-800'; // 완료된 예약
-      }
-      if (reservation.confirmed > 0 && reservation.pending > 0) {
-        return 'bg-blue-300'; // 승인 및 대기 중 예약
-      }
-      if (reservation.confirmed > 0) {
-        return 'bg-orange-200'; // 승인된 예약
-      }
-      if (reservation.pending > 0) {
-        return 'bg-blue-300'; // 대기 중인 예약
-      }
-    }
-    return '';
-  };
-
   const circleColor = getCircleColor();
   const isClickable = !!reservation;
