+++ src/components/pages/myPage/Schedule/mockReservations.ts
@@ -1,0 +1,22 @@
+import { DailyReservation } from '@/types/reservationTypes';
+
+export const mockReservations: DailyReservation[] = [
+  {
+    date: '2024-08-09',
+    reservations: {
+      completed: 1,
+      confirmed: 0,
+      pending: 0,
+    },
+  },
+  {
+    date: '2024-08-15',
+    reservations: {
+      completed: 0,
+      confirmed: 1,
+      pending: 2,
+    },
+  },
+];
+
+export default mockReservations;
