+++ src/types/reservationTypes.ts
@@ -1,0 +1,8 @@
+export interface DailyReservation {
+  date: string;
+  reservations: {
+    completed: number;
+    confirmed: number;
+    pending: number;
+  };
+}
